+++ FSTRaK/FSTrAk.csproj
@@ -169,7 +169,7 @@
     <Compile Include="Utils\NullToVisibilityConverter.cs" />
     <Compile Include="Utils\PathUtil.cs" />
     <Compile Include="Utils\AircraftResolver.cs" />
-    <Compile Include="Utils\ResourceNameToGeometryConverter.cs" />
+    <Compile Include="Utils\Converters.cs" />
     <Compile Include="Utils\SkyVectorMapTileLayer.cs" />
     <Compile Include="Utils\SkyVectorTileSource.cs" />
     <Compile Include="Utils\TimeUtils.cs" />
