+++ FSTRaK/Utils/NullToVisibilityConverter.cs
@@ -5,16 +5,5 @@ using System.Windows;
 
 namespace FSTRaK.Utils
 {
-    public class NullToVisibilityConverter : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            return value == null ? Visibility.Hidden : Visibility.Visible;
-        }
 
-        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
 }
